pyproject.toml
[tool.poetry]
name = "wind-for-life-project"
version = "0.1.0"
description = ""
authors = ["heytem"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.6"
django-environ = "^0.12.0"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
django-allauth = "^65.4.1"
faker = "^36.1.1"
factory-boy = "^3.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
